Typings cache need not be a map but directly on the project
