+++ src/server/_namespaces/ts.server.ts
@@ -8,7 +8,6 @@ export * from "../utilities.js";
 import * as protocol from "./ts.server.protocol.js";
 export { protocol };
 export * from "../scriptInfo.js";
-export * from "../typingsCache.js";
 export * from "../project.js";
 export * from "../editorServices.js";
 export * from "../moduleSpecifierCache.js";
