+++ src/server/types.ts
@@ -1,10 +1,19 @@
 import {
+    ApplyCodeActionCommandResult,
     DirectoryWatcherCallback,
     FileWatcher,
     FileWatcherCallback,
+    InstallPackageOptions,
+    Path,
+    SortedReadonlyArray,
     System,
+    TypeAcquisition,
     WatchOptions,
 } from "./_namespaces/ts.js";
+import {
+    Project,
+    ProjectService,
+} from "./_namespaces/ts.server.js";
 
 export interface CompressedData {
     length: number;
@@ -29,4 +38,20 @@ export interface ServerHost extends System {
     require?(initialPath: string, moduleName: string): ModuleImportResult;
     /** @internal */
     importPlugin?(root: string, moduleName: string): Promise<ModuleImportResult>;
+}
+
+export interface InstallPackageOptionsWithProject extends InstallPackageOptions {
+    projectName: string;
+    projectRootPath: Path;
+}
+
+// for backwards-compatibility
+// eslint-disable-next-line @typescript-eslint/naming-convention
+export interface ITypingsInstaller {
+    isKnownTypesPackageName(name: string): boolean;
+    installPackage(options: InstallPackageOptionsWithProject): Promise<ApplyCodeActionCommandResult>;
+    enqueueInstallTypingsRequest(p: Project, typeAcquisition: TypeAcquisition, unresolvedImports: SortedReadonlyArray<string> | undefined): void;
+    attach(projectService: ProjectService): void;
+    onProjectClosed(p: Project): void;
+    readonly globalTypingsCacheLocation: string | undefined;
 }
